dashboard_streamlit/requirements.txt
streamlit==1.36.0
pandas==2.2.1
numpy==1.26.4
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
plotly==5.22.0
python-dotenv==1.0.1
